Rename packages/builder/src/stores/portal/apps.ts to packages/builder/src/stores/portal/workspaces.ts
