+++ packages/builder/src/settings/pages/auditLogs/index.svelte
@@ -19,7 +19,7 @@
   } from "@budibase/bbui"
   import { licensing } from "@/stores/portal/licensing"
   import { users } from "@/stores/portal/users"
-  import { appsStore } from "@/stores/portal/apps"
+  import { appsStore } from "@/stores/portal/workspaces"
   import { auditLogs } from "@/stores/portal/auditLogs"
   import LockedFeature from "@/pages/builder/_components/LockedFeature.svelte"
   import { createPaginationStore } from "@/helpers/pagination"
