+++ packages/builder/src/settings/pages/people/users/user.svelte
@@ -22,7 +22,7 @@
   import { users } from "@/stores/portal/users"
   import { auth } from "@/stores/portal/auth"
   import { groups } from "@/stores/portal/groups"
-  import { appsStore } from "@/stores/portal/apps"
+  import { appsStore } from "@/stores/portal/workspaces"
   import { licensing } from "@/stores/portal/licensing"
   import { roles } from "@/stores/builder"
   import ForceResetPasswordModal from "./_components/ForceResetPasswordModal.svelte"
