+++ packages/builder/src/settings/routes.ts
@@ -6,7 +6,7 @@ import { Target, type Route } from "@/types/routing"
 import { Pages } from "./pages"
 import { AdminState } from "@/stores/portal/admin"
 import { AppMetaState } from "@/stores/builder/app"
-import { PortalWorkspacesStore } from "@/stores/portal/apps"
+import { PortalWorkspacesStore } from "@/stores/portal/workspaces"
 import { StoreApp } from "@/types"
 import { aiConfigsStore } from "@/stores/portal"
 import { get } from "svelte/store"
