+++ packages/builder/src/stores/builder/deployment.ts
@@ -3,7 +3,7 @@ import { API } from "@/api"
 import { notifications } from "@budibase/bbui"
 import { DeploymentProgressResponse, DeploymentStatus } from "@budibase/types"
 import analytics, { Events, EventSource } from "@/analytics"
-import { appsStore } from "@/stores/portal/apps"
+import { appsStore } from "@/stores/portal/workspaces"
 import { DerivedBudiStore } from "@/stores/BudiStore"
 import { appStore } from "./app"
 import { processStringSync } from "@budibase/string-templates"
