+++ packages/builder/src/stores/portal/index.ts
@@ -3,7 +3,7 @@ import { writable } from "svelte/store"
 export { organisation } from "./organisation"
 export { users } from "./users"
 export { admin } from "./admin"
-export { appsStore, enrichedApps, sortBy } from "./apps"
+export { appsStore, enrichedApps, sortBy } from "./workspaces"
 export { email } from "./email"
 export { auth } from "./auth"
 export { oidc } from "./oidc"
